Fix calc_to_skip func in admin service -> fix get all users bug. Changed check_permissions func

+++ src/services/rest/main_service.py
@@ -14,45 +14,34 @@ is_places_loaded = False
 class MainService:
     def __init__(self):
         self.client = MongoClient('localhost', 27017)
-        # username = os.environ.get('mongo_username') password = os.environ.get('mongo_password') self.client =
-        # MongoClient(f'mongodb+srv://{username}:{password}@cluster.p8ymxwu.mongodb.net/?retryWrites=true&w'
-        # '=majority')
+        #username = os.environ.get('mongo_username')
+        #password = os.environ.get('mongo_password')
+        #self.client = MongoClient(f'mongodb+srv://{username}:{password}@cluster.p8ymxwu.mongodb.net/?retryWrites=true&w'
+        #                           '=majority')
         self.db = self.client.wastewise
         # Create a 2dsphere index on the "data.location.coordinates" field
         self.db.objects.create_index([("data.location.coordinates", "2dsphere")])
 
         global is_places_loaded
         to_load = os.environ.get('Load_Places')
-        if to_load is None:
-            to_load = 'false'
-        if is_places_loaded is False and to_load.lower() in ('true', '1'):
+        if to_load is not None and is_places_loaded is False and to_load.lower() in ('true', '1'):
             # Delete all places from database
             self.db.objects.delete_many({'type': 'place'})
             load_from_shapefile = LoadFromShapefile(self.db)
             load_from_shapefile.load()
             is_places_loaded = True
-        '''to_load = os.environ.get('Load_Places')
-        if to_load == "True" and not is_places_loaded:
-            # Delete all places from database
-            self.db.objects.delete_many({'type': 'place'})
-            # Load places from Excel file
-            load_from_excel = LoadFromExcel(self.db)
-            load_from_excel.load_places()
-            is_places_loaded = True'''
 
     def get_db(self):
         return self.db
 
-    def check_permissions(self, _type: EnumRole, _mail: str, _password: str):
+    def check_permissions(self, _role_name: EnumRole, _mail: str, _password: str, object_type='') -> None:
         user = self.db.users.find_one({'email': _mail})
         if user is None:
             abort(make_response(jsonify(message="There is no user with this email"), HTTPStatus.NOT_FOUND))
         if not pbkdf2_sha256.verify(_password, user['password']):
             abort(make_response(jsonify(message="Wrong Password"), HTTPStatus.UNAUTHORIZED))
-        # if self.db.users.find_one({'email': _mail, 'password': _password}) is None:
-        #    abort(make_response(jsonify(message="Wrong Password"), HTTPStatus.UNAUTHORIZED))
         # check permissions
-        if self.db.users.find_one({'email': _mail, 'role': _type.name}) is None:
+        if object_type != 'PUBLIC_FACILITY' and self.db.users.find_one({'email': _mail, 'role': _role_name.name}) is None:
             abort(make_response(jsonify(message="User doesn't have permissions"), HTTPStatus.FORBIDDEN))
 
     @staticmethod
